(refactor)exporter: update package, function names and comments
